% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el.R
\name{get_external_links}
\alias{get_external_links}
\alias{external_links}
\alias{link_titles}
\alias{link_urls}
\title{Extract External Links}
\usage{
get_external_links(wt)

link_titles(external_links, text = FALSE)

link_urls(external_links, text = FALSE)
}
\arguments{
\item{wt}{a wikitext object returned from
\code{\link{parse_wikitext}}}

\item{external_links}{a list of external links retrieved with
\code{get_external_links}}

\item{text}{whether to return the results
as a character vector. FALSE by default.}
}
\description{
Wikimarkup contains external links,
stored as a URL and a display title, separated
by a space, with the whole thing wrapped in a pair of
square brackets.
These functions allow you to extract external URLs and their
various attributes from wikimarkup.
}
\examples{
\dontrun{

# Get wikitext
wt <- "this is [http://an.external link]"
parsed <- parse_wikitext(wt)

# Extract external link
wls <- get_external_links(parsed)

# Get the actual links
paths <- link_urls(wls)

# Get the display titles
display_title <- link_titles(wls)
}
}
\seealso{
\code{\link{templates}}, \code{\link{headings}} and
\code{\link{wikilinks}} for extractors for other types of
wikimarkup.
}
