% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dest.R
\name{get_dest}
\alias{get_dest}
\title{Get the destination of a link}
\usage{
get_dest(nodes, as_text = FALSE)
}
\description{
For Wikilink and ExternalLink nodes,
\code{get_dest} will retrieve the destination they
are pointing to.
}
