% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{get_displaytext}
\alias{get_displaytext}
\title{Retrieve Displayed Text From Nodes}
\usage{
get_displaytext(nodes, as_text = FALSE)
}
\arguments{
\item{nodes}{a list of nodes, retrieved with \code{\link{get_nodes}}.}

\item{as_text}{whether to return the visible element of each node as
text rather than a Wikimarkup object. FALSE by default.}
}
\value{
a list of either wikimarkup objects or strings, depending
on the value of \code{as_text}. For nodes without a certain
display element (for example, templates), \code{NA}
will be returned.
}
\description{
Some wikimarkup elements (wikilinks or text,
say) have chunks visible when parsed into HTML by MediaWiki.
\code{get_displaytext} retrieves the visible element
for each node in a list of nodes.
}
\examples{
\dontrun{
library(magrittr)
library(mwparser)
raw_wikitext <- "==this is a header=="

text <- read_wikitext(raw_wikitext) \%>\%
   get_nodes(types = "headings") \%>\%
   get_displaytext

text
}
}
\seealso{
\code{\link{get_dest}} to retrieve the link destination
in the case of internal or external links.
}
