% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{get_nodes}
\alias{get_nodes}
\title{Extract Nodes from Wikitext}
\usage{
get_nodes(wikitext, types = c("comments", "external_links", "headings",
  "html", "tags", "templates", "text", "wikilinks"), recursive = TRUE)
}
\arguments{
\item{wikitext}{a chunk of wikitext read in with
\code{\link{read_wikitext}}.}

\item{types}{which type(s) of node to read in. You
can pass in whichever number of supported types you
want.}
}
\description{
\code{get_nodes} retrieves nodes
of various (user-specified) types from a blob of
read-in wikitext. These are returned as a list
of node objects, which can be passed on to
other functions to retrieve components thereof.
}
\examples{
\dontrun{
text <- "this [[is|some]] {{wikitext}}"

parsed_text <- read_wikitext(text)

some_nodes <- get_nodes(parsed_text, types = c("wikilinks", "templates"))
}
}
