% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{read_wikitext}
\alias{read_wikitext}
\title{Read Wikitext}
\usage{
read_wikitext(wikitext, skip_styles = FALSE)
}
\arguments{
\item{wikitext}{a pile of wikitext, as a string, grabbed via
\code{\link{get_wikitext}} or through any other way}

\item{skip_styles}{whether to skip style tags, which (very
occasionally) cause parsing problems. \code{FALSE} by default.}
}
\description{
Initially reads and parses a pile of wikitext,
turning it into an object that can have elements extracted from it.
}
\examples{
# Simple example
wt <- "==this is a header==\\n with {{a|template}}"
parsed <- read_wikitext("wt")

}
