% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{get_headings}
\alias{get_headings}
\alias{templates}
\alias{template_names}
\alias{template_params}
\title{Extract Templates}
\usage{
get_headings(wt, recursive = TRUE)

template_names(templates, text = FALSE)

template_params(templates, text = FALSE)
}
\arguments{
\item{wt}{a wikitext object returned from
\code{\link{parse_wikitext}}}

\item{recursive}{whether to recursively extract
templates. TRUE by default.}

\item{templates}{a list of templates retrieved with
\code{get_templates}}

\item{text}{whether to return the template parameters
as character vectors. FALSE by default.}
}
\description{
Wikimarkup contains templates, stored as
a name and a variable number of parameters, separated
by pipes, with the whole thing wrapped in curly braces.
These functions allow you to extract templates and their
various attributes from wikimarkup.
}
\examples{

# Get wikitext
wt <- "this is {{a|template|okay?}}"
parsed <- parse_wikitext(wt)

# Extract template
templates <- get_templates(parsed)

# Get the templates' names
temp_names <- template_names(templates)

# Get the actual parameter values
parameter_vals <- template_params(templates)

}
\seealso{
\code{\link{headers}}, \code{\link{external_links}} and
\code{\link{wikilinks}} for extractors for other types of
wikimarkup.
}
