% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikilinks.R
\name{get_wikilinks}
\alias{get_wikilinks}
\alias{wikilinks}
\alias{wikilink_paths}
\alias{wikilink_text}
\title{Extract Wikilinks}
\usage{
get_wikilinks(wt)

wikilink_paths(wikilinks, text = FALSE)

wikilink_text(wikilinks, text = FALSE)
}
\arguments{
\item{wt}{a wikitext object returned from
\code{\link{parse_wikitext}}}

\item{wikilinks}{a list of wikilinks retrieved with
\code{get_wikilinks}}

\item{text}{whether to return the results
as a character vector. FALSE by default.}

\item{recursive}{whether to recursively extract
wikilinks TRUE by default.}
}
\description{
Wikimarkup contains internal (wiki) links,
stored as a page name and a display title, separated
by pipes, with the whole thing wrapped in two pairs of
square brackets.
These functions allow you to extract wikilinks and their
various attributes from wikimarkup.
}
\examples{

# Get wikitext
wt <- "this is [[a|link]]"
parsed <- parse_wikitext(wt)

# Extract wikilink
wls <- get_wikilinks(parsed)

# Get the link paths
paths <- wikilink_paths(wls)

# Get the display text
display_text <- wikilink_text(wls)

}
