% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headings.R, R/templates.R
\name{get_headings}
\alias{get_headings}
\alias{headings}
\alias{heading_level}
\alias{heading_title}
\alias{heading_level}
\alias{get_templates}
\title{Extract Headings}
\usage{
get_headings(wt, recursive = TRUE)

heading_level(headings)

heading_title(headings, text = FALSE)

heading_level(headings)

get_templates(wt, recursive = TRUE)
}
\arguments{
\item{wt}{a wikitext object returned from
\code{\link{parse_wikitext}}}

\item{recursive}{whether to recursively extract
headings. TRUE by default.}

\item{headings}{a list of headings retrieved with
\code{get_headings}}

\item{text}{whether to return the heading titles
as text. FALSE by default.}
}
\description{
Wikimarkup contains headers, stored as
text surrounded by a varying number of equals signs
that denote the 'level' of that heading. These
functions allow you to extract headings and their
various attributes from wikimarkup.
}
\examples{

# Get wikitext
wt <- "==this is a header==\\n with {{a|template}}"
parsed <- parse_wikitext(wt)

# Extract headings
headings <- get_headings(parsed)

# Get their levels
levels <- heading_level(headings)

# Get the actual title text
title_text <- heading_title(headings)

}
